Removed igw and vpc_gateway_attachment from template
because it's only used in base_network.
I kept the other ones because they are frequently used by other stacks.

I'm going to remove the dead code in the next comment
